% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_dataset.R
\name{generate_sample_dataset}
\alias{generate_sample_dataset}
\title{Generate and Save Sample Dataset}
\usage{
generate_sample_dataset(n = 50, file_name = "sample_dataset.sav", dir = "data")
}
\arguments{
\item{n}{Number of observations to generate (default is 50)}

\item{file_name}{Name of the file to save (default is "sample_dataset.sav")}

\item{dir}{Directory to save the file (default is "data")}
}
\value{
Invisible NULL. The function is called for its side effect of creating a file.
}
\description{
This function creates a small sample dataset and saves it as a .sav file
in the specified directory.
}
\examples{
\dontrun{
generate_sample_dataset()
generate_sample_dataset(n = 100, file_name = "larger_sample.sav")
}
}
