% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_regions.R
\name{get_country_regions}
\alias{get_country_regions}
\title{Get Region Information for a Specific Country}
\usage{
get_country_regions(x, country_col, region_col, country_code)
}
\arguments{
\item{x}{A data frame containing the country and region information.}

\item{country_col}{The name of the column in 'x' that contains country codes.}

\item{region_col}{The name of the column in 'x' that contains region codes.}

\item{country_code}{The specific country code to analyze.}
}
\value{
A data frame with columns:
  \item{region_code}{The unique region codes for the specified country}
  \item{region_name}{The names of the regions (if available)}
  \item{freq}{The number of occurrences of each region in the dataset}
}
\description{
This function extracts region information for a specified country from a dataset.
It returns a data frame with region codes, names, and counts.
}
\details{
The function expects the region column to be a numeric haven_labelled variable.
It will throw an error if the column is of a different class.
}
