% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_countrycode_scheme.R
\name{infer_countrycode_scheme}
\alias{infer_countrycode_scheme}
\title{Infer Country Code Scheme}
\usage{
infer_countrycode_scheme(x)
}
\arguments{
\item{x}{A character vector of country codes.}
}
\value{
A character string representing the inferred country code scheme.
}
\description{
This function attempts to infer the country code scheme used in a vector of country codes.
}
\details{
The function checks the input against known country code schemes in the countrycode package.
It first looks for a perfect match where all input codes are found in a single scheme.
If no perfect match is found, it returns the scheme with the most matches.
If multiple schemes have the same number of matches, it returns the first one.
}
\examples{
# ISO2 country codes
x <- c("US", "CA", "MX", "FR", "DE")
infer_countrycode_scheme(x)

# ISO3 country codes
y <- c("USA", "CAN", "MEX", "FRA", "DEU")
infer_countrycode_scheme(y)

}
