% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{Load a dataset from an SPSS file}
\usage{
load_dataset(dataset_name, to_csv = TRUE, dataset_dir = NULL)
}
\arguments{
\item{dataset_name}{The name of the dataset (without file extension)}

\item{to_csv}{Logical; if TRUE, saves the dataset as a CSV file}

\item{dataset_dir}{Optional; the directory containing the dataset}
}
\value{
A data frame containing the loaded dataset
}
\description{
This function loads a dataset from an SPSS file and optionally saves it as a CSV.
}
