% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_high_frequency_labels.R
\name{tabulate_high_frequency_labels}
\alias{tabulate_high_frequency_labels}
\title{Tabulate High Frequency Labels}
\usage{
tabulate_high_frequency_labels(x, threshold, verbose = TRUE)
}
\arguments{
\item{x}{A vector of labels to be tabulated.}

\item{threshold}{Numeric. If >= 1, it's treated as an absolute count. If < 1, it's treated as a proportion of non-NA values in x.}

\item{verbose}{Logical. If TRUE (default), prints summary statistics to the console.}
}
\value{
A named integer vector of label frequencies, sorted in descending order, including only labels that meet or exceed the threshold.
}
\description{
This function creates a frequency table of labels in a vector, filtered by a specified threshold.
}
\details{
The function first determines the threshold value. If the input threshold is less than 1, it's interpreted as a proportion
of non-NA values in x. It then creates a frequency table, filters it based on the threshold, and sorts it in descending order.
If verbose is TRUE, it prints summary statistics including the number of labels above the threshold and the number of
participants (non-NA values) represented by these labels.
}
\examples{
x <- c("A", "B", "A", "C", "B", "D", "A", "B", NA)
tabulate_high_frequency_labels(x, threshold = 2)
tabulate_high_frequency_labels(x, threshold = 0.3, verbose = FALSE)

}
