% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_country_mapping.R
\name{construct_country_mapping}
\alias{construct_country_mapping}
\title{Construct Country Mapping}
\usage{
construct_country_mapping(x, alpha_code = "iso3c", warn = TRUE)
}
\arguments{
\item{x}{A character vector of country codes.}

\item{alpha_code}{Character. The desired alpha code format for the output. Default is "iso3c".}

\item{warn}{Logical. Whether to show warnings when country codes cannot be matched. Default is TRUE.}
}
\value{
A data frame with columns:
  \item{Country}{Full country name}
  \item{Code}{Numeric code (0 to n-1)}
  \item{Alpha_Code}{Country code in the specified alpha_code format}
  \item{Source_Var}{Original country code from the input}
  \item{Freq}{Frequency of each country code in the input vector}
}
\description{
This function creates a data frame mapping country codes to names, numeric codes, and frequencies.
}
\details{
The function first infers the country code scheme of the input using `infer_countrycode_scheme`.
It then uses the countrycode package to convert between different country code formats.
Frequencies of each country code in the original input are calculated and included.
}
\examples{
x <- c("USA", "CAN", "MEX", "USA", "CAN", "FRA", "DEU", "USA")
construct_country_mapping(x)

}
